Issue: NET-4277 amxp_expr_buildf with 2 arguments only works in container, not on board [fix]
Problem:
When building a path using amxp_expr_buildf with two arguments
(example: amxp_expr_buildf(&expr, "%s %s", arg1, arg2)),
this works fine in container, but on the board the first argument is used twice
(so as if amxp_expr_buildf(&expr, "%s %s", arg1, arg1) was called).

Cause:
API usage of va_arg was violated (but happens to work on some systems)
"If [the va_list] is passed to a function that uses va_arg(ap,type), then
 the value of ap is undefined after the return of that function." (man va_arg(3))

Solution:
Since we cannot pass va_list multiple times if it's used, and we cannot
get the nth elemnt without hardcoding the types of the previous element,
handle all items of the va_list in one function.
GitOrigin-RevId: 05875da728f0579ebea58affd6ba390f76decca3